pyproject.toml
[project]
name = "bilibili-cli"
version = "0.1.0"
description = "A CLI for Bilibili — browse videos, users, favorites from the terminal"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [{ name = "jackwener", email = "jakevingoo@gmail.com" }]
keywords = ["bilibili", "cli", "terminal", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "bilibili-api-python>=16.0",
    "click>=8.0",
    "rich>=13.0",
    "aiohttp>=3.0",
    "python-dotenv>=1.0",
    "browser-cookie3>=0.19",
]

[project.urls]
Homepage = "https://github.com/jackwener/bilibili-cli"
Repository = "https://github.com/jackwener/bilibili-cli"
Issues = "https://github.com/jackwener/bilibili-cli/issues"

[project.scripts]
bili = "bili_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bili_cli"]
